+++ src/api/domain/tester.api.ts
@@ -1,5 +1,5 @@
-import publicApi from '~/lib/api/config/publicApi';
-import { REQUEST_API } from '~/lib/api/config/requestUrl';
+import publicApi from '~/api/config/publicApi';
+import { REQUEST_API } from '~/api/config/requestUrl';
 
 const testerApi = {
   getTester: () => publicApi.get<{ tester: { id: string; nickname: string } }>(REQUEST_API.TESTER),
