+++ src/app/accounts/callApiTest.tsx
@@ -2,7 +2,7 @@
 
 import { useQuery } from '@tanstack/react-query';
 
-import privateApi from '~/lib/api/config/privateApi';
+import privateApi from '~/api/config/privateApi';
 
 const getCommunityIdCard = async () => privateApi.get('/communities/1/idCards');
 const CallApiTest = () => {
