+++ src/hooks/api/community.query.ts
@@ -1,7 +1,7 @@
 import { useInfiniteQuery } from '@tanstack/react-query';
 
+import communityApi from '~/api/domain/community.api';
 import { community } from '~/hooks/api/queryKey.type';
-import communityApi from '~/lib/api/domain/community.api';
 import { ComunityIdCardsRequest } from '~/types/community/request.type';
 
 export const useGetCommunityIdCards = ({ id, pageParam }: ComunityIdCardsRequest) => {
