+++ src/hooks/api/tester.query.ts
@@ -1,8 +1,8 @@
 import { useQuery } from '@tanstack/react-query';
 import { AxiosError } from 'axios';
 
-import { ErrorResponseType } from '~/lib/api/config/api.types';
-import testerApi from '~/lib/api/domain/tester.api';
+import { ErrorResponseType } from '~/api/config/api.types';
+import testerApi from '~/api/domain/tester.api';
 
 export const useGetTester = () =>
   useQuery<{ tester: { id: string; nickname: string } }, AxiosError<ErrorResponseType>>({
