+++ src/mocks/community/community.mockHandler.ts
@@ -1,6 +1,6 @@
 import { rest } from 'msw';
 
-import { REQUEST_API, ROOT_API_URL } from '~/lib/api/config/requestUrl';
+import { REQUEST_API, ROOT_API_URL } from '~/api/config/requestUrl';
 import { createCommunityDetail, createCommunityIdCards } from '~/mocks/community/community.mock';
 
 const communityMockHandler = [
